requirements.txt

# Core AI & Transformers
transformers>=4.35.0
sentencepiece>=0.1.99
accelerate>=0.24.0
safetensors>=0.4.0
huggingface-hub>=0.20.0

# Stable Diffusion
diffusers>=0.25.0
torch>=2.1.0
scipy>=1.11.0

# PDF & DOCX processing
python-docx>=1.1.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
pdf2image>=1.16.3
opencv-python-headless>=4.8.0.76
Pillow>=10.1.0

# Data & Statistics
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0
seaborn>=0.13.0
matplotlib>=3.8.0
plotly>=5.18.0

# Optional: FAISS (за векторна памет)
faiss-cpu>=1.7.4

# CLI & Utilities
tqdm>=4.66.0
rich>=13.7.0
